+++ src/services/multi-user-accounts/multi-user-accounts.ts
@@ -14,11 +14,16 @@ export interface User {
 }
 
 export interface InviteResponse {
-  email: string
   id: number
   created_at: string
   expires_at: string
-  existing_user_details: User
+  can_be_accepted: boolean
+  is_account_valid: boolean
+  is_invite_active: boolean
+  can_user_join: boolean
+  // These fields leak user information and are excluded entirely for the public endpoint
+  existing_user_details?: User
+  email?: string
 }
 
 export interface UsersMultiAccountResponse {
@@ -27,7 +32,7 @@ export interface UsersMultiAccountResponse {
   last_cancelled_multi_user_account: MultiUserAccountResponse
   is_active_primary: boolean
   is_active_sub: boolean
-  active_invite: InviteResponse
+  active_invites: InviteResponse[]
 }
 
 export interface MultiUserAccountResponse {
@@ -35,13 +40,15 @@ export interface MultiUserAccountResponse {
   product_name: string
   is_active: boolean
   primary_user: User
-  active_invited_emails: string[]
-  available_seats: number
-  available_invites: number
   total_seats: number
-  active_subs: User[]
   end_time: string
   is_primary_account_holder: boolean
+  // The following fields are not included for public or subaccount users
+  active_invited_emails?: InviteResponse[]
+  available_seats?: number
+  available_invites?: InviteResponse[]
+  active_subs?: User[]
+  show_welcome?: boolean
 }
 
 export interface CreateAccountParams {
@@ -51,6 +58,10 @@ export interface CreateAccountParams {
 export interface CreateInvitesParams {
   multi_user_account_id: number
   emails: string[]
+}
+
+export interface UpdateMultiUserAccountParams {
+  show_welcome: boolean
 }
 
 
@@ -76,6 +87,19 @@ export async function createAccount(params: CreateAccountParams): Promise<MultiU
 export async function fetchUsersMultiAccountDetails(userId: number): Promise<UsersMultiAccountResponse> {
   const httpClient = new HttpClient(globalConfig.baseUrl)
   return httpClient.get<UsersMultiAccountResponse>(`${baseUrl}/${userId}/details`)
+}
+
+
+/**
+ * Fetch invite details
+ *
+ * @param {number} inviteId - The ID of the invite to check
+ * @returns {Promise<InviteResponse>} - A promise that resolves to the invite details.
+ * @throws {HttpError} - If the HTTP request fails.
+ */
+export async function fetchInvite(inviteId: number): Promise<InviteResponse> {
+  const httpClient = new HttpClient(globalConfig.baseUrl)
+  return httpClient.get<InviteResponse>(`${baseUrl}/invites/${inviteId}`)
 }
 
 /**
@@ -124,4 +148,16 @@ export async function rescindInvite(inviteId: number): Promise<void> {
  */
 export async function removeUserFromActiveMultiUserAccount(userId: number): Promise<MultiUserAccountResponse|void> {
   return DELETE(`${globalConfig.baseUrl}${baseUrl}/${userId}/remove`, {})
+}
+
+/**
+ * Updates specified fields on a multi-user account. Authorized user must be the primary account owner
+ *
+ * @param {UpdateMultiUserAccountParams} params - The parameters for updating the account.
+ * @returns {Promise<MultiUserAccountResponse>} - Updated MultiUserAccountResponse if account owner
+ * @throws {HttpError} - If the request fails.
+ */
+export async function updateMultiUserAccount(params: UpdateMultiUserAccountParams): Promise<MultiUserAccountResponse> {
+  const httpClient = new HttpClient(globalConfig.baseUrl)
+  return httpClient.patch(`${globalConfig.baseUrl}${baseUrl}/update`, params)
 }